Added custom errors
The Errors module added here is where are the custom errors that can be
raised are stored as well as their messages. These errors don't do
anything special, they just exist as a clean way to express what error
has occurred without a lengthy message.

On branch master
Changes to be committed:
	new file:   lib/errors.rb