Added rubocop configuration
This is the configuration file for rubocop. At current it simply mutes
messages about documentations (I'm lazy and will take care of this
later) and forcing rubocop to look for LF only instead of system default
which is CRLF on Windows. This is done for consistency across OSes.

On branch master
Changes to be committed:
	new file:   .rubocop.yml