Removed duplicated mark and erase methods
Originally, these methods were added as a means to have a shorthand for
dealing with the pencil marks of individual cells. However, since
they're just methods that take a row, column, and value and then call
the related function on the cell it's just as easy to just use the
indexing to find the cell and call the mark and erase methods directly.

On branch master
Changes to be committed:
	modified:   lib/board.rb