Added equality testing to Cell class
Added a .equal? and .== method to Sudoku::Cell that compares pencil
marks and values. It does not determine whether they are equal based on
their status as a constant. Also adds a public .marks method to Cell
that returns a clone of the internal @marks instance variable so that it
cannot be accidentally modified by modifying the returned value.

On branch master
Changes to be committed:
	modified:   lib/cell.rb