Added each_with_index method to Board class
Added a convenience method to iterate over each cell, passing the cell,
the row index, and the column index to the block. This method helps
prevent code that needs to iterate each cell on the board from needing
the deeper levels of nesting encapsulated by the method.

On branch master
Changes to be committed:
	modified:   lib/board.rb