Added a parameter to cell value set for constants
Realized how annoying it is that you couldn't set a value that isn't a
pencil mark without it being automatically set as a constant. This
change adds a parameter that defaults to false that will set the
constant value, thus allowing "final" marks without necessarily making
the field a constant.

On branch master
Changes to be committed:
	modified:   lib/cell.rb