Fixed bug with value setting in Cells
Introducing the ability to choose whether or not to set a cell as a
constant after updating it's value also introduced a bug. The change
was set in an assignment function, which doesn't work. The fix was to
add a new set_value function that accepted this new parameter and change
the assignment function to use this instead.

On branch master
Changes to be committed:
	modified:   lib/cell.rb